+++ spec/whois/record/parser/responses/whois.markmonitor.com/status_registered_spec.rb
@@ -21,9 +21,19 @@ describe Whois::Record::Parser::WhoisMarkmonitorCom, "status_registered.expected
     described_class.new(part)
   end
 
+  describe "#domain" do
+    it do
+      expect(subject.domain).to eq("google.com")
+    end
+  end
+  describe "#domain_id" do
+    it do
+      expect(subject.domain_id).to eq("")
+    end
+  end
   describe "#status" do
     it do
-      expect { subject.status }.to raise_error(Whois::AttributeNotSupported)
+      expect(subject.status).to eq(:registered)
     end
   end
   describe "#available?" do
@@ -39,27 +49,27 @@ describe Whois::Record::Parser::WhoisMarkmonitorCom, "status_registered.expected
   describe "#created_on" do
     it do
       expect(subject.created_on).to be_a(Time)
-      expect(subject.created_on).to eq(Time.parse("1999-03-15"))
+      expect(subject.created_on).to eq(Time.parse("2002-10-02 00:00:00 -0700"))
     end
   end
   describe "#updated_on" do
     it do
       expect(subject.updated_on).to be_a(Time)
-      expect(subject.updated_on).to eq(Time.parse("2011-02-11"))
+      expect(subject.updated_on).to eq(Time.parse("2013-10-29 11:50:06 -0700"))
     end
   end
   describe "#expires_on" do
     it do
       expect(subject.expires_on).to be_a(Time)
-      expect(subject.expires_on).to eq(Time.parse("2012-03-14"))
+      expect(subject.expires_on).to eq(Time.parse("2020-09-13 21:00:00 -0700"))
     end
   end
   describe "#registrar" do
     it do
       expect(subject.registrar).to be_a(Whois::Record::Registrar)
-      expect(subject.registrar.id).to eq(nil)
-      expect(subject.registrar.name).to eq("Markmonitor.com")
-      expect(subject.registrar.organization).to eq(nil)
+      expect(subject.registrar.id).to eq("292")
+      expect(subject.registrar.name).to eq("MarkMonitor, Inc.")
+      expect(subject.registrar.organization).to eq("MarkMonitor, Inc.")
       expect(subject.registrar.url).to eq("http://www.markmonitor.com")
     end
   end
@@ -69,14 +79,14 @@ describe Whois::Record::Parser::WhoisMarkmonitorCom, "status_registered.expected
       expect(subject.registrant_contacts).to have(1).items
       expect(subject.registrant_contacts[0]).to be_a(Whois::Record::Contact)
       expect(subject.registrant_contacts[0].type).to eq(Whois::Record::Contact::TYPE_REGISTRANT)
-      expect(subject.registrant_contacts[0].name).to eq("DNS Admin")
+      expect(subject.registrant_contacts[0].name).to eq("Dns Admin")
       expect(subject.registrant_contacts[0].organization).to eq("Google Inc.")
-      expect(subject.registrant_contacts[0].address).to eq("1600 Amphitheatre Parkway")
+      expect(subject.registrant_contacts[0].address).to eq("Please contact contact-admin@google.com, 1600 Amphitheatre Parkway")
       expect(subject.registrant_contacts[0].city).to eq("Mountain View")
       expect(subject.registrant_contacts[0].zip).to eq("94043")
       expect(subject.registrant_contacts[0].state).to eq("CA")
       expect(subject.registrant_contacts[0].country_code).to eq("US")
-      expect(subject.registrant_contacts[0].phone).to eq("+1.6506234000")
+      expect(subject.registrant_contacts[0].phone).to eq("+1.6502530000")
       expect(subject.registrant_contacts[0].fax).to eq("+1.6506188571")
       expect(subject.registrant_contacts[0].email).to eq("dns-admin@google.com")
       expect(subject.registrant_contacts[0].created_on).to eq(nil)
@@ -111,13 +121,13 @@ describe Whois::Record::Parser::WhoisMarkmonitorCom, "status_registered.expected
       expect(subject.technical_contacts[0].type).to eq(Whois::Record::Contact::TYPE_TECHNICAL)
       expect(subject.technical_contacts[0].name).to eq("DNS Admin")
       expect(subject.technical_contacts[0].organization).to eq("Google Inc.")
-      expect(subject.technical_contacts[0].address).to eq("1600 Amphitheatre Parkway")
+      expect(subject.technical_contacts[0].address).to eq("2400 E. Bayshore Pkwy")
       expect(subject.technical_contacts[0].city).to eq("Mountain View")
       expect(subject.technical_contacts[0].zip).to eq("94043")
       expect(subject.technical_contacts[0].state).to eq("CA")
       expect(subject.technical_contacts[0].country_code).to eq("US")
-      expect(subject.technical_contacts[0].phone).to eq("+1.6506234000")
-      expect(subject.technical_contacts[0].fax).to eq("+1.6506188571")
+      expect(subject.technical_contacts[0].phone).to eq("+1.6503300100")
+      expect(subject.technical_contacts[0].fax).to eq("+1.6506181499")
       expect(subject.technical_contacts[0].email).to eq("dns-admin@google.com")
       expect(subject.technical_contacts[0].created_on).to eq(nil)
       expect(subject.technical_contacts[0].updated_on).to eq(nil)
@@ -129,12 +139,20 @@ describe Whois::Record::Parser::WhoisMarkmonitorCom, "status_registered.expected
       expect(subject.nameservers).to have(4).items
       expect(subject.nameservers[0]).to be_a(Whois::Record::Nameserver)
       expect(subject.nameservers[0].name).to eq("ns2.google.com")
+      expect(subject.nameservers[0].ipv4).to eq(nil)
+      expect(subject.nameservers[0].ipv6).to eq(nil)
       expect(subject.nameservers[1]).to be_a(Whois::Record::Nameserver)
       expect(subject.nameservers[1].name).to eq("ns1.google.com")
+      expect(subject.nameservers[1].ipv4).to eq(nil)
+      expect(subject.nameservers[1].ipv6).to eq(nil)
       expect(subject.nameservers[2]).to be_a(Whois::Record::Nameserver)
-      expect(subject.nameservers[2].name).to eq("ns4.google.com")
+      expect(subject.nameservers[2].name).to eq("ns3.google.com")
+      expect(subject.nameservers[2].ipv4).to eq(nil)
+      expect(subject.nameservers[2].ipv6).to eq(nil)
       expect(subject.nameservers[3]).to be_a(Whois::Record::Nameserver)
-      expect(subject.nameservers[3].name).to eq("ns3.google.com")
+      expect(subject.nameservers[3].name).to eq("ns4.google.com")
+      expect(subject.nameservers[3].ipv4).to eq(nil)
+      expect(subject.nameservers[3].ipv6).to eq(nil)
     end
   end
 end